backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.22.0
torch>=2.0.0
transformers>=4.30.0
faiss-cpu>=1.9.0
numpy>=1.21.0
sentence-transformers>=2.2.2
